# src/back/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.CreateGradeSheetRequest:
    properties:
      course_id:
        type: integer
      description:
        type: string
      sheet_url:
        type: string
    required:
    - course_id
    - description
    - sheet_url
    type: object
...
  handlers.UpdateGradeSheetRequest:
    properties:
      description:
        type: string
      sheet_url:
        type: string
    required:
    - description
    - sheet_url
    type: object
...
  models.GradeSheet:
    properties:
      course_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      sheet_url:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8443
info:
  contact:
    email: support@laritmo.local
    name: API Support
  description: Educational portal API with course management
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Laritmo API
  version: "1.0"
paths:
...
  /api/admin/grade-sheets:
    post:
      consumes:
      - application/json
      description: Create a new grade sheet (admin only)
      parameters:
      - description: Grade sheet data
        in: body
        name: gradesheet
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateGradeSheetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.GradeSheet'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create grade sheet
      tags:
      - admin-gradesheets
  /api/admin/grade-sheets/{id}:
    delete:
      description: Delete grade sheet by ID (admin only)
      parameters:
      - description: Grade Sheet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete grade sheet
      tags:
      - admin-gradesheets
    put:
      consumes:
      - application/json
      description: Update grade sheet by ID (admin only)
      parameters:
      - description: Grade Sheet ID
        in: path
        name: id
        required: true
        type: integer
      - description: Grade sheet data
        in: body
        name: gradesheet
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateGradeSheetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update grade sheet
      tags:
      - admin-gradesheets
...
  /api/grade-sheets:
    get:
      description: Get list of all grade sheets with optional course filter
      parameters:
      - description: Course ID filter
        in: query
        name: course_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.GradeSheet'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all grade sheets
      tags:
      - grade-sheets
  /api/grade-sheets/{id}:
    get:
      description: Get grade sheet details by ID
      parameters:
      - description: Grade Sheet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GradeSheet'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get grade sheet by ID
      tags:
      - grade-sheets
...
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"